crawling/requirements.txt
beautifulsoup4==4.5.3
click==6.7
lxml==3.7.3
pysolr==3.6.0
python-dotenv==0.6.4
requests==2.13.0
watson-developer-cloud==0.25.2
elasticsearch>=5.0.0,<6.0.0
tinydb